Adds primary adapter - BatchNoteKeeperHandler
