+++ patterns/hexagonal-architecture/src/main/java/com/baeldung/notekeeper/core/port/notifier/NoteNotifier.java
@@ -1,4 +1,4 @@
-package com.baeldung.notekeeper.core.notifier;
+package com.baeldung.notekeeper.core.port.notifier;
 
 import com.baeldung.notekeeper.core.model.Note;
